+++ TPV2/src/game/Characters/Togo/Togo.h
@@ -5,28 +5,28 @@ class DinoShield;
 class Togo : public Character
 {
 public:
-	Togo(FightManager* mngr, b2Vec2 pos, char input,int p);
+	Togo(FightManager* mngr, b2Vec2 pos, char input, ushort p);
 	virtual ~Togo();
 
-	virtual void BasicNeutral(int frameNumber) override;
-	virtual void BasicForward(int frameNumber) override;
-	virtual void BasicUpward(int frameNumber) override;
-	virtual void BasicDownward(int frameNumber) override;
+	virtual void BasicNeutral(ushort frameNumber) override;
+	virtual void BasicForward(ushort frameNumber) override;
+	virtual void BasicUpward(ushort frameNumber) override;
+	virtual void BasicDownward(ushort frameNumber) override;
 
-	virtual void SpecialNeutral(int frameNumber) override;
-	virtual void SpecialForward(int frameNumber) override;
-	virtual void SpecialUpward(int frameNumber) override;
-	virtual void SpecialDownward(int frameNumber) override;
+	virtual void SpecialNeutral(ushort frameNumber) override;
+	virtual void SpecialForward(ushort frameNumber) override;
+	virtual void SpecialUpward(ushort frameNumber) override;
+	virtual void SpecialDownward(ushort frameNumber) override;
 
 	virtual bool GetHit(HitData a, Entity* attacker, bool& controlHitLag, bool& controlShake, bool& controlCamShake) override;
 	virtual void update() override;
 	virtual void SetSpear(bool spear);
-	void SpecialLHit(int frameNumber);
+	void SpecialLHit(ushort frameNumber);
 	virtual void draw(SDL_Rect* camera)override;
 private:
 	bool lanza = true;
 	float spearTime;
-	int attackBase;
+	ushort attackBase;
 	float attackMul;
 	DinoShield* dShield = nullptr;
 	SDL_Rect bite;
